+++ src/components/CheckboxGroup/CheckboxGroup.tsx
@@ -162,14 +162,13 @@ export const CheckboxGroup = chakra(
             {newChildren}
           </Stack>
         </ChakraCheckboxGroup>
-        {footnote && showHelperInvalidText && (
-          <HelperErrorText
-            id={`${id}-helperErrorText`}
-            isInvalid={isInvalid}
-            text={footnote}
-            __css={styles.helperErrorText}
-          />
-        )}
+        <HelperErrorText
+          id={`${id}-helperErrorText`}
+          isInvalid={isInvalid}
+          renderText={showHelperInvalidText}
+          text={footnote}
+          __css={styles.helperErrorText}
+        />
       </Fieldset>
     );
   })